feat: protecting endpoints, security tuning and new activity column

+++ src/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service'
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -37,14 +38,14 @@ export class TasksService {
 
   }
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const newTask = await this.prisma.task.create({
         data: {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
-          userId: createTaskDto.userId
+          userId: tokenPayload.sub
         }
       })
 
@@ -55,7 +56,7 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
+  async update(id: number, updateTaskDto: UpdateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -64,6 +65,10 @@ export class TasksService {
       })
 
       if (!findTask) {
+        throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
         throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
       }
 
@@ -86,7 +91,7 @@ export class TasksService {
   }
 
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -96,6 +101,10 @@ export class TasksService {
 
       if (!findTask) {
         throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException("Falha ao deletar essa tarefa!", HttpStatus.BAD_REQUEST)
       }
 
       await this.prisma.task.delete({
